LayoutTests/fast/dom/HTMLElement/attr-dir-auto-change-before-text-node.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
<script src="../../js/resources/js-test-pre.js"></script>
<style>
.testDiv {-webkit-border-start: 5px solid green; -webkit-border-end: 5px solid red; }
</style>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>

<div id="test" dir="auto" class="testDiv"></div>
<script>

description('Test that directionality is re-evaluated when an element with dir=auto set, gets a new node before the node that was used to determine its directionality.');

var el = document.getElementById("test");
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(255, 0, 0)'");
var child1 = document.createElement("div");
child1.innerHTML = "מקור השם עברית";
el.appendChild(child1);
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(0, 128, 0)'");

var child2 = document.createElement("div");
child2.innerHTML = "test";
el.insertBefore(child2, child1);
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(255, 0, 0)'");

el.removeChild(child2);
shouldBe("document.defaultView.getComputedStyle(el, null).getPropertyValue('border-right-color')", "'rgb(0, 128, 0)'");

el.style.display = "none";
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
